@@ -1,153 +1,148 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.Graph;
-namespace LeaderAnalytics.Vyntix.Web.Model
+﻿namespace LeaderAnalytics.Vyntix.Web.Model;
+
+public class UserRecord // Named to not conflict with Graph.User
 {
-    public class UserRecord // Named to not conflict with Graph.User
+    private string _email;
+    public User User { get; set; }
+
+    public string BillingID
     {
-        private string _email;
-        public User User { get; set; }
-
-        public string BillingID 
+        get
         {
-            get 
-            {
-                if (User.AdditionalData.TryGetValue(UserAttributes.BillingID, out object obj))
-                    return obj?.ToString();
+            if (User.AdditionalData.TryGetValue(UserAttributes.BillingID, out object obj))
+                return obj?.ToString();
 
-                return null;
-            }
-            set { User.AdditionalData[UserAttributes.BillingID] = value; } 
+            return null;
         }
-        
-        public bool IsBanned 
-        { 
-            get 
-            {
-                if (User.AdditionalData.TryGetValue(UserAttributes.IsBanned, out object obj))
-                    return Convert.ToBoolean(obj);
-                
-                return false;
-            } 
-            
-            set { User.AdditionalData[UserAttributes.IsBanned] = value; } 
-        }
-        
-        public bool IsCorporateAdmin 
-        { 
-            get
-            {
-                if (User.AdditionalData.TryGetValue(UserAttributes.IsCorporateAdmin, out object obj))
-                    return Convert.ToBoolean(obj);
-                
-                return false;
-            }
+        set { User.AdditionalData[UserAttributes.BillingID] = value; }
+    }
 
-            set { User.AdditionalData[UserAttributes.IsCorporateAdmin] = value; }
+    public bool IsBanned
+    {
+        get
+        {
+            if (User.AdditionalData.TryGetValue(UserAttributes.IsBanned, out object obj))
+                return Convert.ToBoolean(obj);
+
+            return false;
         }
 
-        public bool IsOptIn 
+        set { User.AdditionalData[UserAttributes.IsBanned] = value; }
+    }
+
+    public bool IsCorporateAdmin
+    {
+        get
         {
-            get
-            {
-                if (User.AdditionalData.TryGetValue(UserAttributes.IsOptIn, out object obj))
-                    return Convert.ToBoolean(obj);
-                
-                return false;
-            }
-            set
-            {
-                User.AdditionalData[UserAttributes.IsOptIn] = value;
-            }
+            if (User.AdditionalData.TryGetValue(UserAttributes.IsCorporateAdmin, out object obj))
+                return Convert.ToBoolean(obj);
+
+            return false;
         }
 
-        public string PaymentProviderCustomerID 
+        set { User.AdditionalData[UserAttributes.IsCorporateAdmin] = value; }
+    }
+
+    public bool IsOptIn
+    {
+        get
         {
-            get
-            {
-                if (User.AdditionalData.TryGetValue(UserAttributes.PaymentProviderCustomerID, out object obj))
-                    return obj?.ToString();
-                
-                return null;
-            }
-            set
-            {
-                User.AdditionalData[UserAttributes.PaymentProviderCustomerID] = value;
-            }
+            if (User.AdditionalData.TryGetValue(UserAttributes.IsOptIn, out object obj))
+                return Convert.ToBoolean(obj);
+
+            return false;
         }
-        
-        public DateTime? SuspendedUntil 
+        set
         {
-            get
-            {
-                if (User.AdditionalData.TryGetValue(UserAttributes.SuspendedUntil, out object obj))
-                    return obj == null ? (DateTime?)null : Convert.ToDateTime(obj.ToString().Substring(1)).ToUniversalTime();
-                
-                return null;
-            }
-            set
-            {
-                User.AdditionalData[UserAttributes.SuspendedUntil] = value.HasValue ? "z" + value.Value.ToString("o") : null;
-            }
-        } // UTC
+            User.AdditionalData[UserAttributes.IsOptIn] = value;
+        }
+    }
 
-        public string EMailAddress
+    public string PaymentProviderCustomerID
+    {
+        get
         {
-            get {
+            if (User.AdditionalData.TryGetValue(UserAttributes.PaymentProviderCustomerID, out object obj))
+                return obj?.ToString();
 
-                if (!string.IsNullOrEmpty(_email))
-                    return _email;
+            return null;
+        }
+        set
+        {
+            User.AdditionalData[UserAttributes.PaymentProviderCustomerID] = value;
+        }
+    }
 
-                if ((User?.Identities ?? null) != null)
-                    _email = User.Identities.FirstOrDefault(x => x.SignInType == "emailAddress")?.IssuerAssignedId;
-                
-                if (string.IsNullOrEmpty(_email))
-                    _email = User?.OtherMails?.FirstOrDefault();
+    public DateTime? SuspendedUntil
+    {
+        get
+        {
+            if (User.AdditionalData.TryGetValue(UserAttributes.SuspendedUntil, out object obj))
+                return obj == null ? (DateTime?)null : Convert.ToDateTime(obj.ToString().Substring(1)).ToUniversalTime();
 
+            return null;
+        }
+        set
+        {
+            User.AdditionalData[UserAttributes.SuspendedUntil] = value.HasValue ? "z" + value.Value.ToString("o") : null;
+        }
+    } // UTC
+
+    public string EMailAddress
+    {
+        get
+        {
+
+            if (!string.IsNullOrEmpty(_email))
                 return _email;
-            }
-        }
 
-        public bool IsLocalAccount
+            if ((User?.Identities ?? null) != null)
+                _email = User.Identities.FirstOrDefault(x => x.SignInType == "emailAddress")?.IssuerAssignedId;
+
+            if (string.IsNullOrEmpty(_email))
+                _email = User?.OtherMails?.FirstOrDefault();
+
+            return _email;
+        }
+    }
+
+    public bool IsLocalAccount
+    {
+        // https://docs.microsoft.com/en-us/graph/api/resources/user?view=graph-rest-1.0
+        //
+        // creationType String  
+        //
+        // Indicates whether the user account was created as a regular school or work account(null), 
+        // an external account(Invitation), 
+        // a local account for an Azure Active Directory B2C tenant(LocalAccount) 
+        // or self - service sign - up using email verification (EmailVerified). Read - only.
+
+        // 
+        get
         {
-            // https://docs.microsoft.com/en-us/graph/api/resources/user?view=graph-rest-1.0
-            //
-            // creationType String  
-            //
-            // Indicates whether the user account was created as a regular school or work account(null), 
-            // an external account(Invitation), 
-            // a local account for an Azure Active Directory B2C tenant(LocalAccount) 
-            // or self - service sign - up using email verification (EmailVerified). Read - only.
-
-            // 
-            get
-            {
-                return User.CreationType == "LocalAccount";
-            }
+            return User.CreationType == "LocalAccount";
         }
+    }
 
-        public UserRecord()
+    public UserRecord()
+    {
+        User = new User
         {
-            User = new User
-            {
-                PasswordProfile = new PasswordProfile(),
-                Identities = new List<ObjectIdentity>(5),
-                OtherMails = new List<string>(5),
-                AdditionalData = new Dictionary<string, object>()
-            };
-        }
+            PasswordProfile = new PasswordProfile(),
+            Identities = new List<ObjectIdentity>(5),
+            OtherMails = new List<string>(5),
+            AdditionalData = new Dictionary<string, object>()
+        };
+    }
 
-        public UserRecord(User user)
-        {
-            if (user == null)
-                throw new ArgumentNullException("user");
+    public UserRecord(User user)
+    {
+        if (user == null)
+            throw new ArgumentNullException("user");
 
-            User = user;
+        User = user;
 
-            if (User.AdditionalData == null)
-                User.AdditionalData = new Dictionary<string, object>();
-        }
+        if (User.AdditionalData == null)
+            User.AdditionalData = new Dictionary<string, object>();
     }
 }
