@@ -1,68 +1,57 @@
-﻿using LeaderAnalytics.Vyntix.Web.Model;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Text;
-using Autofac;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿namespace LeaderAnalytics.Vyntix.Web.Tests;
 
-namespace LeaderAnalytics.Vyntix.Web.Tests
+public class BaseTest
 {
-    public class BaseTest
+    protected GraphClientCredentials GraphClientCredentials { get; private set; }
+    protected ContainerBuilder ContainerBuilder { get; set; }
+    protected IContainer Container { get; set; }
+    protected IConfiguration Configuration { get; set; }
+    protected IServiceCollection ServiceCollection { get; set; }
+    protected ServiceRegistrar registrar { get; set; }
+
+
+    public BaseTest()
     {
-        protected GraphClientCredentials GraphClientCredentials { get; private set; }
-        protected ContainerBuilder ContainerBuilder { get; set; }
-        protected IContainer Container { get; set; }
-        protected IConfiguration Configuration { get; set; }
-        protected IServiceCollection ServiceCollection { get; set; }
-        protected ServiceRegistrar registrar { get; set; }
+        ServiceCollection = new ServiceCollection();
+        IConfiguration config = new ConfigurationBuilder().AddJsonFile("appsettings.json").Build();
+        registrar = new ServiceRegistrar(config, "Development");
+        registrar.ConfigureServices(ServiceCollection);
+    }
 
-
-        public BaseTest()
+    protected GraphClientCredentials GetGraphCredentials()
+    {
+        IConfigurationSection graphSection = Configuration.GetSection("AzureGraph");
+        GraphClientCredentials = new GraphClientCredentials
         {
-            ServiceCollection = new ServiceCollection();
-            IConfiguration config = new ConfigurationBuilder().AddJsonFile("appsettings.json").Build();
-            registrar = new ServiceRegistrar(config, "Development");
-            registrar.ConfigureServices(ServiceCollection);
-        }
+            ClientID = graphSection.GetValue<string>("ClientID"),
+            TenantID = graphSection.GetValue<string>("TenantID"),
+            ClientSecret = graphSection.GetValue<string>("ClientSecret")
+        };
 
-        protected GraphClientCredentials GetGraphCredentials()
-        {
-            IConfigurationSection graphSection = Configuration.GetSection("AzureGraph");
-            GraphClientCredentials = new GraphClientCredentials
-            {
-                ClientID = graphSection.GetValue<string>("ClientID"),
-                TenantID = graphSection.GetValue<string>("TenantID"),
-                ClientSecret = graphSection.GetValue<string>("ClientSecret")
-            };
+        return GraphClientCredentials;
+    }
 
-            return GraphClientCredentials;
-        }
-        
-        protected virtual void Setup()
-        {
-            CreateContainer();
-            CreateMocks();
-            Container = ContainerBuilder.Build();
-        }
+    protected virtual void Setup()
+    {
+        CreateContainer();
+        CreateMocks();
+        Container = ContainerBuilder.Build();
+    }
 
 
-        protected virtual void CreateMocks()
-        {
-            
-        }
+    protected virtual void CreateMocks()
+    {
 
-        protected virtual void CreateContainer()
-        {
-            ContainerBuilder = new ContainerBuilder();
-            new AutofacModule().ConfigureServices(ContainerBuilder, registrar);
-        }
+    }
 
-        protected ILifetimeScope GetLifetimeScope()
-        {
-            return Container.BeginLifetimeScope();
-        }
+    protected virtual void CreateContainer()
+    {
+        ContainerBuilder = new ContainerBuilder();
+        new AutofacModule().ConfigureServices(ContainerBuilder, registrar);
+    }
+
+    protected ILifetimeScope GetLifetimeScope()
+    {
+        return Container.BeginLifetimeScope();
     }
 }
