@@ -1,53 +1,46 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace LeaderAnalytics.Vyntix.Web.Domain;
 
-namespace LeaderAnalytics.Vyntix.Web.Domain
+public class AsyncResult
 {
-    public class AsyncResult
+    public bool Success { get; set; }
+    public string ErrorMessage { get; set; }
+}
+
+public class AsyncResult<T> : AsyncResult
+{
+    public T Result { get; set; }
+    public int TotalResultCount { get; set; }
+}
+
+public class RowOpResult
+{
+    public enum RowOp
     {
-        public bool Success { get; set; }
-        public string ErrorMessage { get; set; }
+        Fail,
+        Insert,
+        Update,
+        Delete
     }
 
-    public class AsyncResult<T> : AsyncResult
+    public RowOp Result { get; set; } = RowOp.Fail;
+    public string ErrorMessage { get; set; }
+}
+
+public class RowOpResult<T> : RowOpResult
+{
+    public T Item { get; set; }
+
+    public RowOpResult() { }
+
+    public RowOpResult(T item)
     {
-        public T Result { get; set; }
-        public int TotalResultCount { get; set; }
+        this.Item = item;
     }
+}
 
-    public class RowOpResult
-    {
-        public enum RowOp
-        {
-            Fail,
-            Insert,
-            Update,
-            Delete
-        }
-
-        public RowOp Result { get; set; } = RowOp.Fail;
-        public string ErrorMessage { get; set; }
-    }
-
-    public class RowOpResult<T> : RowOpResult
-    {
-        public T Item { get; set; }
-
-        public RowOpResult() { }
-
-        public RowOpResult(T item)
-        {
-            this.Item = item;
-        }
-    }
-
-    public class ValidationResult<T>
-    {
-        public bool Success { get; set; }
-        public string ErrorMessage { get; set; }
-        public T ExistingEntity { get; set; }
-    }
+public class ValidationResult<T>
+{
+    public bool Success { get; set; }
+    public string ErrorMessage { get; set; }
+    public T ExistingEntity { get; set; }
 }
