@@ -1,12 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿namespace LeaderAnalytics.Vyntix.Web.Domain;
 
-namespace LeaderAnalytics.Vyntix.Web.Domain
+public interface IValueParameter<T>
 {
-    public interface IValueParameter<T> 
-    {
-        T Value { get; set; }
-    }
+    T Value { get; set; }
 }
