@@ -1,16 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿namespace LeaderAnalytics.Vyntix.Web.Model;
 
-namespace LeaderAnalytics.Vyntix.Web.Model
+public class EmailMessage
 {
-    public class EmailMessage
-    {
-        public string[] To { get; set; }
-        public string From { get; set; }
-        public string Subject { get; set; }
-        public string Msg { get; set; }
-        public bool IsHTML { get; set; }
-    }
+    public string[] To { get; set; }
+    public string From { get; set; }
+    public string Subject { get; set; }
+    public string Msg { get; set; }
+    public bool IsHTML { get; set; }
 }
