@@ -1,14 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿namespace LeaderAnalytics.Vyntix.Web.Model;
 
-namespace LeaderAnalytics.Vyntix.Web.Model
+public class GraphClientCredentials
 {
-    public class GraphClientCredentials
-    {
-        public string TenantID { get; set; }
-        public string ClientID { get; set; }
-        public string ClientSecret { get; set; }
-    }
+    public string TenantID { get; set; }
+    public string ClientID { get; set; }
+    public string ClientSecret { get; set; }
 }
